requirements.txt
# pip install -r requirements.txt

click==6.7
cycler==0.11.0
fbpca==1.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.0.0
Flask-SQLAlchemy==2.4.1
fonttools==4.31.1
itsdangerous==0.24
Jinja2==2.11.2
joblib==1.1.0
kiwisolver==1.4.0
MarkupSafe==1.1.1
matplotlib==3.5.1
numpy==1.21.5
packaging==21.3
pandas==1.3.5
Pillow==9.0.1
PyJWT==1.7.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
SQLAlchemy==1.3.6
threadpoolctl==3.1.0
typing_extensions==4.1.1
Werkzeug==1.0.1
gunicorn

tensorflow_gpu==2.3.0
